+++ src/ai_labeler/ai.py
@@ -1,88 +1,7 @@
-import os
-from pathlib import Path
-import yaml
-from enum import Enum
 import controlflow as cf
 from typing import Union
-from dataclasses import dataclass
 from .github import PullRequest, Issue, Label
-
-
-@dataclass
-class LabelConfig:
-    name: str
-    description: str | None = None
-    instructions: str | None = None
-
-
-@dataclass
-class Config:
-    instructions: str
-    include_repo_labels: bool
-    labels: list[LabelConfig]
-    context_files: list[str] = None  # New field
-
-    @classmethod
-    def load(cls) -> "Config":
-        # Get config path from action input, falling back to default
-        workspace_path = os.getenv("GITHUB_WORKSPACE", "")
-        config_path = os.getenv("INPUT_CONFIG-PATH", ".github/ai-labeler.yml")
-        full_path = Path(workspace_path) / config_path
-
-        try:
-            with open(full_path) as f:
-                data = yaml.safe_load(f)
-
-            labels_data = data.get("labels", [])
-            label_configs = []
-
-            for item in labels_data:
-                if isinstance(item, str):
-                    # Simple string label
-                    label_configs.append(LabelConfig(name=item))
-                else:
-                    # Dict with label name as key
-                    # item is like {"label_name": {"description": "...", "instructions": "..."}}
-                    name, props = next(iter(item.items()))
-                    if props is None:
-                        props = {}
-                    label_configs.append(
-                        LabelConfig(
-                            name=name,
-                            description=props.get("description"),
-                            instructions=props.get("instructions"),
-                        )
-                    )
-
-            return cls(
-                instructions=data.get("instructions", ""),
-                include_repo_labels=data.get("include_repo_labels", True),
-                labels=label_configs,
-                context_files=data.get("context_files", []),
-            )
-        except FileNotFoundError:
-            # If no config file exists, return default config
-            return cls(
-                instructions="",
-                include_repo_labels=True,
-                labels=[],
-                context_files=[],
-            )
-
-    def load_context_files(self, workspace_path: str = None) -> dict[str, str]:
-        """Load the contents of context files"""
-        if not workspace_path:
-            workspace_path = os.getenv("GITHUB_WORKSPACE", "")
-
-        context = {}
-        for file_path in self.context_files or []:
-            full_path = Path(workspace_path) / file_path
-            try:
-                with open(full_path) as f:
-                    context[file_path] = f.read()
-            except FileNotFoundError:
-                print(f"Warning: Context file {file_path} not found")
-        return context
+from .label_workflow import Config
 
 
 @cf.flow
